🚀 Feature: Add volume input range with changing icons
